Fix critical code quality issues and add validation
Critical fixes:
- Add division by zero check in Vector3D::operator/
- Add temperature range validation in Species thermodynamic functions
- Add negative diffusion coefficient check in FickDiffusion

Medium priority fixes:
- Add NASA polynomial coefficient size validation
- Add matrix index bounds checking in MatrixXd
- Change ISolver iteration types from int to size_t

Build fixes:
- Add missing headers (stdexcept, string, functional)
- Update TestSolver to match ISolver interface changes

All fixes tested and verified with successful build and unit tests.

Files modified:
- core/include/core/types/CommonTypes.h
- chemistry/include/chemistry/species/Species.h
- physics/include/physics/diffusion/FickDiffusion.h
- core/include/core/interfaces/ISolver.h
- core/src/test_interfaces.cpp
- parallel/include/parallel/solver/MPIDiffusionSolver.h
- CODE_FIXES_SUMMARY.md (new)

+++ core/src/test_interfaces.cpp
@@ -48,9 +48,9 @@ public:
     std::string getName() const override { return "TestSolver"; }
     void setTolerance(double /*tol*/) override {}
     double getTolerance() const override { return 1e-6; }
-    void setMaxIterations(int /*maxIter*/) override {}
-    int getMaxIterations() const override { return 100; }
-    int getIterationCount() const override { return 0; }
+    void setMaxIterations(size_t /*maxIter*/) override {}
+    size_t getMaxIterations() const override { return 100; }
+    size_t getIterationCount() const override { return 0; }
     double getResidual() const override { return 0.0; }
     void reset() override {}
     size_t getNumDOFs() const override { return 0; }
